creado nextion serial cediquim
