+++ minute_purge.json
@@ -1,0 +1,1 @@
+{"min":2}
