+++ upload-nextion.sh
@@ -1,0 +1,1 @@
+python3 nextion-upload-rpi.py rpi-cediquim-nextion.tft /dev/ttyS0 NX4827T043
